Pass all tests for exercise 4
Tested using loop over rest parameters
to get the index of items to remove,
then use splice to remove the item from the array
and mutate the original array. Was able to pass
test one with the function using those methods but
could not pass test 2. Had ideas for conditionals
to use for some of the later tests, but did not
move on from
test 2.

Researched other methods to utilize such as
find(), map(), and join(). Read up on use cases of
each before trying to use sort to create a new
array that does not include the items to be
removed.