Remove older code comments
Removed older code that was being used to try
to solve test 2 using indexOf() and splice().

This was the exercise that took me the longest so
far but I did not give up and was eventually able
to pass the test scenarios by researching several
concepts and testing my code with them.

Rest will be useful for variadic functions
where mutliple arguments need passed, and seems
simpler to use than using the older arguments
syntax.