Erase existing code and begin new approach to solution
Was originally solving the problem with solution
that would determine which number was greater,
then loop from the smaller number adding
each subsequent number up to the larger number
using an assignment operator.

Had trouble solving with this approach, returned
to outlining the problem on paper and work out the
math further. After working on an outline of whats
needed to complete the math I came across Gauss
Summation to calculate total from 1 to n. Going
to adapt this to calculate total sum from any(x)
number to n.