Fix value validation to confirm number is positive integer
Corrected syntax for my typeof check that was
breaking the test from initiating. After fixing
that my code is passing all of the tests conditions
but not returning as passed ie. the conditions that
should alert error do read back as error but alert
not appearing as passed for the unit test.